# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScreenIQ Background Screening Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates/{candidateId}/run-screening:
    post:
      operationId: runScreening
      tags: [screening]
      summary: Trigger background screening for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreeningInput"
      responses:
        "201":
          description: Screening initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningRun"
  /candidates/{candidateId}/screening-runs:
    get:
      operationId: listScreeningRuns
      tags: [screening]
      summary: List screening runs for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScreeningRun"
  /candidates/{candidateId}/screening-runs/{runId}:
    get:
      operationId: getScreeningRun
      tags: [screening]
      summary: Get a specific screening run
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
        - name: runId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningRun"
  /candidates/{candidateId}/screening-runs/{runId}/checks:
    get:
      operationId: listCheckResults
      tags: [screening]
      summary: List all check results for a screening run
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
        - name: runId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckResult"

  # ── Risk Assessment ──────────────────────────────────────────
  /candidates/{candidateId}/screening-runs/{runId}/risk-assessment:
    get:
      operationId: getRiskAssessment
      tags: [risk]
      summary: Get AI risk assessment for a screening run
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
        - name: runId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAssessment"
        "404":
          description: Not yet generated
    post:
      operationId: generateRiskAssessment
      tags: [risk]
      summary: Generate AI risk assessment for a completed screening run
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
        - name: runId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Risk assessment generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAssessment"

  # ── Adverse Action ───────────────────────────────────────────
...
  /candidates/{candidateId}/disputes:
    get:
      operationId: listCandidateDisputes
      tags: [disputes]
      summary: List disputes filed by a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dispute"
    post:
      operationId: createDispute
      tags: [disputes]
      summary: File a dispute for a check result
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "201":
          description: Dispute filed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /disputes/{id}:
    get:
      operationId: getDispute
      tags: [disputes]
      summary: Get a dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
    patch:
      operationId: updateDispute
      tags: [disputes]
      summary: Update a dispute status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"

  # ── Compliance ───────────────────────────────────────────────
...
components:
  schemas:
...
    ScreeningInput:
      type: object
      properties:
        checkTypes:
          type: array
          items:
            type: string
    ScreeningRun:
      type: object
      required: [id, candidateId, status, checksTotal, checksCompleted, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        status:
          type: string
          enum: [pending, running, completed, failed]
        checksTotal:
          type: integer
        checksCompleted:
          type: integer
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        checks:
          type: array
          items:
            $ref: "#/components/schemas/CheckResult"
    CheckResult:
      type: object
      required: [id, screeningRunId, checkType, status, dataSource, confidenceScore, processingTimeMs, createdAt]
      properties:
        id:
          type: integer
        screeningRunId:
          type: integer
        checkType:
          type: string
          enum: [criminal, employment, education, driving]
        status:
          type: string
          enum: [clear, flag, review, confirmed, unverified, discrepancy, violations, suspended, clean]
        statusLabel:
          type: string
        dataSource:
          type: string
        confidenceScore:
          type: number
        processingTimeMs:
          type: integer
        details:
          type: object
        createdAt:
          type: string

    # ── Risk Assessment ───────────────────────────────────────
    RiskAssessment:
      type: object
      required: [id, screeningRunId, overallRisk, keyFindings, recommendedSteps, fcraAdverseFlag, createdAt]
      properties:
        id:
          type: integer
        screeningRunId:
          type: integer
        overallRisk:
          type: string
          enum: [Low, Medium, High]
        keyFindings:
          type: string
        recommendedSteps:
          type: string
        fcraAdverseFlag:
          type: boolean
        riskFactors:
          type: array
          items:
            type: string
        createdAt:
          type: string

    # ── Adverse Action ────────────────────────────────────────
...
    DisputeInput:
      type: object
      required: [checkResultId, reason]
      properties:
        checkResultId:
          type: integer
        reason:
          type: string
        supportingDocUrl:
          type: string
    DisputeUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, under_review, resolved, rejected]
        resolution:
          type: string
    Dispute:
      type: object
      required: [id, candidateId, checkResultId, reason, status, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        checkResultId:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [open, under_review, resolved, rejected]
        resolution:
          type: ["string", "null"]
        supportingDocUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    # ── Compliance ────────────────────────────────────────────
...