# lib/api-zod/src/generated/types/dispute.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ScreenIQ Background Screening Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DisputeStatus } from './disputeStatus';

export interface Dispute {
  id: number;
  candidateId: number;
  checkResultId: number;
  reason: string;
  status: DisputeStatus;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  supportingDocUrl?: string | null;
  createdAt: string;
  updatedAt?: string;
}