# lib/api-zod/src/generated/types/screeningRun.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ScreenIQ Background Screening Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckResult } from './checkResult';
import type { ScreeningRunStatus } from './screeningRunStatus';

export interface ScreeningRun {
  id: number;
  candidateId: number;
  status: ScreeningRunStatus;
  checksTotal: number;
  checksCompleted: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  checks?: CheckResult[];
}